Move branchwater code into method_sourmash.py
In hindsight the separate module wasn't really needed,
given that branchwater is just a sourmash plugin and
aims to implement the same methods only faster (with
the long term goal of improving the base sourmash).
